refactor: Extract duplicated logic into a middleware function
