+++ dotcom-rendering/src/server/server.dev.ts
@@ -1,11 +1,4 @@
-import {
-	type ErrorRequestHandler,
-	type Handler,
-	type NextFunction,
-	type Request,
-	type Response,
-	Router,
-} from 'express';
+import { type ErrorRequestHandler, type Handler, Router } from 'express';
 import { pages } from '../devServer/routers/pages';
 import { targets } from '../devServer/routers/targets';
 import { handleAllEditorialNewslettersPage } from './handler.allEditorialNewslettersPage.web';
@@ -35,6 +28,7 @@ import {
 import { handleAppsThrasher } from './handler.thrasher.apps';
 import { getABTestsFromQueryParams } from './lib/get-abtests-from-query-params';
 import { getContentFromURLMiddleware } from './lib/get-content-from-url';
+import { responseHeaderMiddleware } from './lib/header-middleware';
 import { requestLoggerMiddleware } from './lib/logging-middleware';
 import { recordError } from './lib/logging-store';
 
@@ -67,8 +61,6 @@ const editionalisefront = (url: string): string => {
 	}
 	return url;
 };
-
-const backendApp = process.env.GU_APP ?? 'dev';
 
 const redirects: Handler = (req, res, next) => {
 	const path = req.path.split('/')[1];
@@ -114,19 +106,7 @@ const renderer = Router();
 renderer.use(getContentFromURLMiddleware);
 renderer.use(getABTestsFromQueryParams);
 renderer.use(requestLoggerMiddleware);
-
-renderer.use((req: Request, res: Response, next: NextFunction) => {
-	const headers = {
-		'X-Gu-Backend-App': backendApp,
-		'X-Gu-Backend-App-Target-Group': 'dev',
-	};
-
-	for (const [key, value] of Object.entries(headers)) {
-		res.setHeader(key, value);
-	}
-
-	next();
-});
+renderer.use(responseHeaderMiddleware);
 
 renderer.get('/Article/*url', handleArticle);
 renderer.get('/Interactive/*url', handleInteractive);
